pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "learning-branch-cut"
description = "Framework for learning-augmented branch-and-cut experiments on repeated MILP instances"
version = "0.0.1"
authors = [{ name = "OptiProject Contributors" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "scikit-learn>=1.3"
]

[project.optional-dependencies]
solver = [
    "pyscipopt>=4.5; platform_system != 'Windows' or platform_machine != 'arm64'"
]
dev = [
    "pytest>=7.4",
    "black>=24.0",
    "ruff>=0.3"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["learning_branch_cut*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
